Removed unneeded textview notification and switched to using textview delegate method. Started working on new screen size code for ios 8. Removed unneeded nil checks from delegate calls. Switch float type of framechanged delegate to be CGFloat. Added default init method that gives a screen width and 54 point height.

+++ MessageComposerView/MessageComposerView.h
@@ -30,7 +30,7 @@
 - (void)messageComposerSendMessageClickedWithMessage:(NSString*)message;
 @optional
 // executed whenever the MessageComposerView's frame changes. Provides the frame it is changing to and the animation duration
-- (void)messageComposerFrameDidChange:(CGRect)frame withAnimationDuration:(float)duration;
+- (void)messageComposerFrameDidChange:(CGRect)frame withAnimationDuration:(CGFloat)duration;
 // executed whenever the user is typing in the text view
 - (void)messageComposerUserTyping;
 @end
